scan ok
Former-commit-id: b2967622cfb33811d99d1b220947aa3a816b0ec0

+++ unisim_simulators/cxx/leonsav/scanner.ops.cc
@@ -44,7 +44,7 @@ Scanner::ISA::~ISA()
 {}
 
 Scanner::Operation*
-Scanner::ISA::decode(uint64_t addr, uint32_t code, std::string& disasm)
+Scanner::ISA::decode(uint32_t addr, uint32_t code, std::string& disasm)
 {
   std::ostringstream buf;
   Scanner::Operation* op = 0;
